requirements.txt
django
django-countries
flake8
mypy
pylint
black
djangorestframework
psycopg2-binary
factory_boy
celery
redis
djoser
djangorestframework-simplejwt
